requirements.txt
flask
flask_sqlalchemy
python-dotenv
fpdf
python-docx
pymysql
openai
# Additional libraries for specific functionalities